[tfds]: Significant fix for TyDiQA (and to a smaller extent XQuAD).
Fix issue with squad-like processing that incorrectly strips
contexts. Specifically, in qa_utils.generate_squadlike_examples() we see a
.strip() call that modifies both titles and contexts. Unfortunately, answer
spans are computed against the _original_ (unstripped) context. By stripping
and not updating `answer_start`s, the answer span of the contexts with leading
white-space become incorrect.

This CL fixes the issue by simply avoiding modifications to the title and
context strings. This CL also makes a small modification to a test example that
trigger the issue when the strip() call is still in place.

Preliminary analysis on TyDi (where I originally found this problem) and XQuAD show:

For TyDi: 1283 out of 5077 in the 'validation' set are impacted (25%)
For TyDi: 13492 out of 49881 in the 'train' set are impacted (27%)
For XQuAD: 34 out of the 240 test paragraphs are impacted (14%)

There are also negligible changes to Squad V1 and V2. A single dev context is
changed and around 19 contexts are affected in train (out of ~18,000-19,000).

This CL does not seem to impact MLQA.

PiperOrigin-RevId: 366153930